Renaming package component to composable.
