+++ shared/src/commonMain/kotlin/me/inassar/App.kt
@@ -5,7 +5,7 @@ import androidx.compose.material.Scaffold
 import androidx.compose.runtime.Composable
 import cafe.adriel.voyager.navigator.CurrentScreen
 import cafe.adriel.voyager.navigator.Navigator
-import me.inassar.common.component.TopBar
+import me.inassar.common.composable.TopBar
 import me.inassar.features.feature.presentation.ui.products.ProductsScreen
 
 @Composable
